requirements.txt
flask
flask_wtf
mysql_connector
black